python/requirements.txt
msrestazure==0.6.4
loguru==0.6.0
pyyaml==6.0
pyopenssl==22.0.0
adal==1.2.7
azure-common==1.1.28
azure-appconfiguration==1.3.0
azure-identity==1.10.0
azure-keyvault==4.2.0
azure-storage-common==2.1.0
azure-mgmt-core==1.3.2
azure-mgmt-resource==21.1.0
azure-mgmt-containerservice==20.3.0
azure-mgmt-keyvault==10.1.0
azure-mgmt-storage==20.1.0
